moved all ranger job abilities to job_utils
