+++ scripts/actions/abilities/barrage.lua
@@ -8,11 +8,11 @@
 local abilityObject = {}
 
 abilityObject.onAbilityCheck = function(player, target, ability)
-    return 0, 0
+	return xi.job_utils.ranger.checkBarrage(player, target, ability)
 end
 
-abilityObject.onUseAbility = function(player, target, ability)
-    player:addStatusEffect(xi.effect.BARRAGE, 0, 0, 60)
+abilityObject.onUseAbility = function(player, target, ability, action)
+	return xi.job_utils.ranger.useBarrage(player, target, ability, action)
 end
 
 return abilityObject
