+++ scripts/actions/abilities/unlimited_shot.lua
@@ -8,11 +8,11 @@
 local abilityObject = {}
 
 abilityObject.onAbilityCheck = function(player, target, ability)
-    return 0, 0
+    return xi.job_utils.ranger.checkUnlimitedShot(player, target, ability)
 end
 
-abilityObject.onUseAbility = function(player, target, ability)
-    player:addStatusEffect(xi.effect.UNLIMITED_SHOT, 1, 0, 60)
+abilityObject.onUseAbility = function(player, target, ability, action)
+	return xi.job_utils.ranger.useUnlimitedShot(player, target, ability, action)
 end
 
 return abilityObject
